Faster feedback on input validation errors
Before, when putting in a custom seq, loaders would show, indicating that inference was starting.
Now, validation happens before loaders are displayed.

I also lowered the timeout that the script waits for typing because now it's more obvious exactly when the model is running inference.

The result is a much snappier-feeling page.