New go parenthood that includes GO:0140323 (#53)
Uzay Sezen reports an issue with running inference on B9HQN5 where no results are returned. [1]

I believe this is an issue with the go parenthood that's used by the model, not the model itself. This is because GO:0140323 was missing from the old parenthood, causing the UI to hang.

This change updates the parenthood file to be consistent (I hope) with the one the model was trained on.
I also stripped whitespace from the json file, which makes it slightly smaller, making a slightly faster page load.

Code to generate the parenthood file:

```
go_terms = parenthood_lib.parse_full_go_file()
go_term_dict = {
    t.term_id: list(label for _, label in t.related_labels)
    for t in go_terms
}
with open('go_parenthood.json', 'wb') as f:
    f.write(json.dumps(go_term_dict, separators=(',', ':')).encode())
```

Confirmed that with the new parenthood, B9HQN5 inference results are reasonable.

Tested for regressions:

- TrpC go listing does not change
- Hemoglobin go listing has one more label, transporter activity, which is a reasonable inclusion. Moreover, this was already in the go chart beforehand (AFAICT the chart and the list use different sources of truth, which should probably be fixed, eventually)
- alpha amylase go listing does not change

[1] https://twitter.com/UzaySezen/status/1513543547621761024